+++ reductstore/src/api/http.rs
@@ -313,7 +313,6 @@ pub(crate) mod tests {
     use reduct_base::error::ReductError as BaseHttpError;
     use reduct_base::ext::ExtSettings;
     use reduct_base::msg::bucket_api::BucketSettings;
-    use reduct_base::msg::lifecycle_api::LifecycleSettings;
     use reduct_base::msg::replication_api::{ReplicationMode, ReplicationSettings};
     use reduct_base::msg::server_api::ServerInfo;
     use reduct_base::msg::token_api::{Permissions, TokenCreateRequest};
@@ -542,91 +541,6 @@ pub(crate) mod tests {
             let keeper = keeper.await;
             let components = keeper.get_anonymous().await.unwrap();
             assert!(components.storage.info().await.is_ok());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_stop_replication_tasks(#[future] keeper: Arc<StateKeeper>) {
-            let keeper = keeper.await;
-            let components = keeper.get_anonymous().await.unwrap();
-
-            {
-                let mut repo = components.replication_repo.write().await.unwrap();
-                repo.start();
-                assert!(repo.is_replication_running("api-test").await.unwrap());
-            }
-
-            keeper.stop_replication_tasks().await.unwrap();
-
-            let components = keeper.get_anonymous().await.unwrap();
-            let repo = components.replication_repo.read().await.unwrap();
-            assert!(!repo.is_replication_running("api-test").await.unwrap());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_stop_lifecycle_tasks(#[future] keeper: Arc<StateKeeper>) {
-            let keeper = keeper.await;
-            let components = keeper.get_anonymous().await.unwrap();
-
-            {
-                let mut repo = components.lifecycle_repo.write().await.unwrap();
-                repo.create_lifecycle(
-                    "api-test",
-                    LifecycleSettings {
-                        bucket: "bucket-1".to_string(),
-                        older_than: "1d".to_string(),
-                        interval: "1h".to_string(),
-                        ..LifecycleSettings::default()
-                    },
-                )
-                .await
-                .unwrap();
-                repo.start().await.unwrap();
-                assert!(repo.is_lifecycle_running("api-test").await.unwrap());
-            }
-
-            keeper.stop_lifecycle_tasks().await.unwrap();
-
-            let components = keeper.get_anonymous().await.unwrap();
-            let repo = components.lifecycle_repo.read().await.unwrap();
-            assert!(!repo.is_lifecycle_running("api-test").await.unwrap());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_sync_storage(#[future] keeper: Arc<StateKeeper>) {
-            let keeper = keeper.await;
-            let components = keeper.get_anonymous().await.unwrap();
-            let bucket = components
-                .storage
-                .get_bucket("bucket-1")
-                .await
-                .unwrap()
-                .upgrade_and_unwrap();
-
-            let mut writer = bucket
-                .begin_write("entry-sync", 1, 4, "text/plain".to_string(), HashMap::new())
-                .await
-                .unwrap();
-            writer.send(Ok(Some(Bytes::from("test")))).await.unwrap();
-            writer.send(Ok(None)).await.unwrap();
-
-            keeper.sync_storage().await.unwrap();
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_stop_replication_tasks_not_ready(
-            #[future] not_ready_keeper: Arc<StateKeeper>,
-        ) {
-            let err = not_ready_keeper
-                .await
-                .stop_replication_tasks()
-                .await
-                .err()
-                .unwrap();
-            assert_eq!(err.status, ErrorCode::ServiceUnavailable);
         }
 
         #[rstest]
